+++ src/components/eteamlist.component.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import axios from "../utils/axios";
-
+import Swal from "sweetalert2";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faAmbulance,
@@ -92,7 +92,19 @@ const ETeam = (props) => {
           <button
             className="btn btn-sm btn-danger"
             onClick={() => {
-              props.deleteETeam(props.eteam.username);
+              Swal.fire({
+                title: "Are you sure?",
+                text: "You won't be able to revert this!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes, delete it!",
+              }).then((result) => {
+                if (result.isConfirmed) {
+                  props.deleteETeam(props.eteam.username);
+                }
+              });
             }}
           >
             <FontAwesomeIcon icon={faTrashAlt} />
@@ -155,6 +167,21 @@ export default class ETeamList extends Component {
       })
       .then((response) => {
         console.log(response.data);
+        if (response.data.success === false) {
+          Swal.fire({
+            icon: "error",
+            title: "Oops...",
+            text: response.data.message,
+          });
+        } else {
+          Swal.fire({
+            position: "center",
+            icon: "success",
+            title: response.data.message,
+            showConfirmButton: false,
+            timer: 2000,
+          });
+        }
       });
 
     this.setState({
@@ -164,6 +191,14 @@ export default class ETeamList extends Component {
   }
 
   async updateETeam(username, name, contactNumber) {
+    if(!name || name.length<5 || !contactNumber || contactNumber.length<3){
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Invalid data",
+      });
+      return;
+    }
 
     await axios
       .post("police/eteam/update/", {
