+++ src/components/eventlist.component.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import axios from "../utils/axios";
-
+import Swal from "sweetalert2";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faExclamationTriangle,
@@ -206,7 +206,19 @@ const Event = (props) => {
           <button
             className="btn btn-sm btn-danger"
             onClick={() => {
-              props.deleteEvent(props.event.id);
+              Swal.fire({
+                title: "Are you sure?",
+                text: "You won't be able to revert this!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes, delete it!",
+              }).then((result) => {
+                if (result.isConfirmed) {
+                  props.deleteEvent(props.event.id);
+                }
+              });
             }}
           >
             <FontAwesomeIcon icon={faTrashAlt} />
@@ -274,6 +286,21 @@ export default class EventList extends Component {
       })
       .then((response) => {
         console.log(response.data);
+        if (response.data.success === false) {
+          Swal.fire({
+            icon: "error",
+            title: "Oops...",
+            text: response.data.message,
+          });
+        } else {
+          Swal.fire({
+            position: "center",
+            icon: "success",
+            title: response.data.message,
+            showConfirmButton: false,
+            timer: 2000,
+          });
+        }
       });
 
     // this.setState({
@@ -292,6 +319,15 @@ export default class EventList extends Component {
     kmPost,
     suburb
   ) {
+    if(!eventDate || !eventTime ||!kmPost || 0>kmPost || kmPost>127){
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Invalid data",
+      });
+      return;
+    }
+
     console.log("Update date format");
     const accDate = new Date(eventDate);
     const datetime = new Date(
