+++ src/components/policelist.component.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import axios from "../utils/axios";
-
+import Swal from "sweetalert2";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faUsers,
@@ -84,7 +84,19 @@ const Police = (props) => {
           <button
             className="btn btn-sm btn-danger"
             onClick={() => {
-              props.deletePolice(props.police.username);
+              Swal.fire({
+                title: "Are you sure?",
+                text: "You won't be able to revert this!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes, delete it!",
+              }).then((result) => {
+                if (result.isConfirmed) {
+                  props.deletePolice(props.police.username);
+                }
+              });
             }}
           >
             <FontAwesomeIcon icon={faTrashAlt} />
@@ -147,6 +159,21 @@ export default class PoliceList extends Component {
       })
       .then((response) => {
         console.log(response.data);
+        if (response.data.success === false) {
+          Swal.fire({
+            icon: "error",
+            title: "Oops...",
+            text: response.data.message,
+          });
+        } else {
+          Swal.fire({
+            position: "center",
+            icon: "success",
+            title: response.data.message,
+            showConfirmButton: false,
+            timer: 2000,
+          });
+        }
       });
 
     this.setState({
@@ -156,6 +183,14 @@ export default class PoliceList extends Component {
   }
 
   async updatePolice(username, name, adminRights) {
+    if(!name || name.length<5){
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Invalid data",
+      });
+      return;
+    }
     await axios
       .post("police/update/", {
         username: username,
